# backend/docs/openapi-gpt-action.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Isekai Lucas Engine GPT Actions",
    "version": "0.22.0",
    "description": "Schema normal para GPT Actions. Expone endpoints de lectura, busqueda, preview y mutacion central controlada. No expone rollback, admin ni mutadores peligrosos de combate/misiones."
  },
  "servers": [
    {
      "url": "https://isekai-lucas-engine.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/characters/{characterId}/state": {
      "get": {
        "operationId": "getCharacterState",
        "summary": "Obtiene estado mecanico directo de un personaje.",
        "description": "Devuelve HP, MP, hambre, energia, dinero, inventario, equipo, skills, condiciones y acumuladores biologicos sin reconstruir desde logs.",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "isekai_lucas_main"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado directo del personaje.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}